feat: buildings page has its own layout #15
