+++ frontend/src/i18n/en/index.ts
@@ -114,6 +114,7 @@ export default {
   nominal_pga_z: 'Nominal PGA Z-dir. [g]',
   open_measured_data: 'Open Measured Data',
   original_size: 'Original',
+  other_experiments_of_reference: 'Other experiments of this reference',
   overview: 'Overview',
   paper: 'Paper',
   pga_axis: 'PGA [g units]',
