+++ frontend/src/router/routes.ts
@@ -2,14 +2,14 @@ import { RouteRecordRaw } from 'vue-router';
 
 const routes: RouteRecordRaw[] = [
   {
+    path: '/buildings',
+    component: () => import('layouts/BuildingsLayout.vue'),
+  },
+  {
     path: '/',
     component: () => import('layouts/MainLayout.vue'),
     children: [
       { path: '', component: () => import('pages/IndexPage.vue') },
-      {
-        path: '/buildings',
-        component: () => import('src/pages/BuildingsPage.vue'),
-      },
       {
         path: '/test/:id',
         component: () => import('src/pages/TestPage.vue'),
